fixed an error that came from putting everything in functions; not sure why

+++ readpdf.sh
@@ -22,8 +22,12 @@ temp_audio=""
 
 # function to clean up temp files
 cleanup() {
-	echo "> cleaning up temporary files"
-  	rm -rf "${temp_dir}"
+	# don't run if program exits early somewhere else
+	if [ ! -z "$temp_dir" ]
+	then
+		echo "> cleaning up temporary files"
+		rm -rf "${temp_dir}"
+	fi
 }
 
 # bind cleanup function to exit, incl. interrupt
@@ -53,131 +57,125 @@ usage() {
 
 }
 
+# check requirements before running
+requirements
+
 # options: -i [input] -v [voice] -o -c -h 
 # i : input- argument, specify input file
 # v : voice- specify voice
 # o : output- flag, not arg; create file with base filename, and write audio to it
 # c : convert- flag, if present, convert to mp3
 # h : help- display usage info
-init() {
-	# check requirements before running
-	requirements
 
-	# parse options and arguments given using getopts
-	while getopts "ohci:v:" opt; do
-		case $opt in
-			h)
-				usage
-				exit 1
-				;;
-			i)
-				if [ -e "$OPTARG" ]
-				then
-					input_file="$OPTARG"
-				else
-					echo "Error: file \"${OPTARG}\" does not exist"
-					exit 1
-				fi
-				;;
-			o)
-				output_file=true
-				;;
-			c)
-				convert=true
-				if [ "$output_file" != true ]
-				then
-					echo "Can\'t convert without output"
-					exit 1
-				fi
-				;;
-			v)
-				# validate voice choice
-				if [ ! -z "$(say -v '?' | grep -i ${OPTARG})" ]
-				then
-					voice="$OPTARG"
-				else
-					echo "Error: voice not found, defaulting to 'Tom'"
-				fi
-				;;
-			*)
-				usage
-				exit 1
-				;;
-		esac
-	done
-}
-
-# execute the meaty instructions
-main() {
-	# initialize
-	init
-
-	# check if the input file is specified and extant
-	if [ "$input_file" != "" ] && [ -e "$input_file" ]
-	then
-		# extract the name of the file and extension
-		name="$(echo "$input_file" | cut -f 1 -d '.')"
-		extension="$(echo "$input_file" | awk -F. '{print $NF}')"
-
-
-		# create temporary files if they are needed inside temp directory
-		temp_dir="$(mktemp -d)"
-		temp_text="${temp_dir}/text.txt)"
-		touch "$temp_text"
-
-		if [ "$convert" = true ]
-		then 
-			temp_audio="$(touch ${temp_dir}/audio.aiff)"
-		fi
-
-		# convert to txt
-		echo "> converting to txt"
-		if [ "$extension" == "pdf" ]
-		then
-			gs -q -dNOPROMPT -dBATCH -dNOPAUSE -sDEVICE=txtwrite -sOutputFile="$temp_text" "$input_file"
-		elif [ "$extension" == "epub" ]
-		then
-			pandoc "$input_file" -f epub -t plain -o "$temp_text"
-		else
-			echo "Unsupported filetype, attempting implicit pandoc conversion"
-
-			# if implicit conversion succeeds, this command will return an empty string
-			# if if does not return a zero type (-z), then exit the script
-			if [ ! -z "$(pandoc ${input_file} -t plain -o ${temp_text})" ]
+# parse options and arguments given using getopts
+while getopts "ohci:v:" opt; do
+	case $opt in
+		h)
+			usage
+			exit 1
+			;;
+		i)
+			
+			if [ ! -e "$OPTARG" ]
 			then
+				echo "Error: file \"${OPTARG}\" does not exist"
 				exit 1
-			fi
-		fi
-
-		# if output is true create one, else do it in place
-		# having trouble using pipes, so txt middle man might be necessary evil
-		if [ "$output_file" = true ]
-		then
-			echo "> generating audio"
-			if [ "$convert" = true ] 
-			then
-				say -v 'Tom' -f "$temp_text" -o "$temp_audio"
-				# get the bitrate of output file and use that for equivalent conversion
-				bit="$(ffmpeg -i "${temp_audio}" 2>&1 | grep Audio | awk -F", " '{print $5}' | cut -d' ' -f1)"
-				echo "> converting audio"
-				ffmpeg -hide_banner -loglevel fatal -i "$temp_audio" -f mp3 -acodec libmp3lame -ab "$bit"k "./${name}.mp3"
 			else
-				say -v 'Tom' -f "$temp_text" -o "./${name}.aiff"
+			input_file="$OPTARG"
 			fi
+			# echo "$input_file"
+			;;
+		o)
+			output_file=true
+			;;
+		c)
+			convert=true
+			if [ "$output_file" != true ]
+			then
+				echo "Can\'t convert without output"
+				exit 1
+			fi
+			;;
+		v)
+			# validate voice choice
+			if [ ! -z "$(say -v '?' | grep -i ${OPTARG})" ]
+			then
+				voice="$OPTARG"
+			else
+				echo "Error: voice not found, defaulting to 'Tom'"
+			fi
+			;;
+		*)
+			usage
+			exit 1
+			;;
+	esac
+done
 
 
-		else
-			echo "> starting narration"
-			say -v 'Tom' -f "$temp_text"
+
+
+# check if the input file is specified and extant
+if [ "$input_file" != "" ] && [ -e "$input_file" ]
+then
+	# extract the name of the file and extension
+	name="$(echo "$input_file" | cut -f 1 -d '.')"
+	extension="$(echo "$input_file" | awk -F. '{print $NF}')"
+
+
+	# create temporary files if they are needed inside temp directory
+	temp_dir="$(mktemp -d)"
+	temp_text="${temp_dir}/text.txt)"
+	touch "$temp_text"
+
+	if [ "$convert" = true ]
+	then 
+		temp_audio="$(touch ${temp_dir}/audio.aiff)"
+	fi
+
+	# convert to txt
+	echo "> converting to txt"
+	if [ "$extension" == "pdf" ]
+	then
+		gs -q -dNOPROMPT -dBATCH -dNOPAUSE -sDEVICE=txtwrite -sOutputFile="$temp_text" "$input_file"
+	elif [ "$extension" == "epub" ]
+	then
+		pandoc "$input_file" -f epub -t plain -o "$temp_text"
+	else
+		echo "Unsupported filetype, attempting implicit pandoc conversion"
+
+		# if implicit conversion succeeds, this command will return an empty string
+		# if if does not return a zero type (-z), then exit the script
+		if [ ! -z "$(pandoc ${input_file} -t plain -o ${temp_text})" ]
+		then
+			exit 1
 		fi
+	fi
+
+	# if output is true create one, else do it in place
+	# having trouble using pipes, so txt middle man might be necessary evil
+	if [ "$output_file" = true ]
+	then
+		echo "> generating audio"
+		if [ "$convert" = true ] 
+		then
+			say -v 'Tom' -f "$temp_text" -o "$temp_audio"
+			# get the bitrate of output file and use that for equivalent conversion
+			bit="$(ffmpeg -i "${temp_audio}" 2>&1 | grep Audio | awk -F", " '{print $5}' | cut -d' ' -f1)"
+			echo "> converting audio"
+			ffmpeg -hide_banner -loglevel fatal -i "$temp_audio" -f mp3 -acodec libmp3lame -ab "$bit"k "./${name}.mp3"
+		else
+			say -v 'Tom' -f "$temp_text" -o "./${name}.aiff"
+		fi
+
 
 	else
-		echo "Error: no input file specified"
-		echo "" 
-		usage
+		echo "> starting narration"
+		say -v 'Tom' -f "$temp_text"
 	fi
-}
 
-
-# the ol one liner
-main
+else
+	echo "Error: no input file specified"
+	echo "$input_file" 
+	usage
+fi
